% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_fordeling.R
\name{forbered_data_fordeling}
\alias{forbered_data_fordeling}
\title{Preprosessering}
\usage{
forbered_data_fordeling(data)
}
\arguments{
\item{data}{datasett "licorice gargle"}
}
\value{
datasett med norske nivåer og verdier
}
\description{
Preprosessering
}
