% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetRegData.R
\name{getRegData}
\alias{getRegData}
\title{Provides a dataframe containing data from a registry}
\usage{
getRegData()
}
\value{
regData data frame
}
\description{
Provides a dataframe containing data from a registry
}
