% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeHist.R
\name{makeHist}
\alias{makeHist}
\title{Make a histogram, either plot or its data}
\usage{
makeHist(df, x, var, makeTable = FALSE)
}
\arguments{
\item{df}{dataframe from which output is to be made}

\item{x}{What to plot}

\item{var}{string defining which variable in the data frame to use}

\item{makeTable}{Logical that if TRUE function will return a data frame
containin the bin borders and count within each bin}
}
\value{
a graphical object or data frame
}
\description{
Short demo on how to produce dynamic content in a shiny app at Rapporteket
}
