% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeHist.R
\name{makeSimpleHist}
\alias{makeSimpleHist}
\title{Make a histogram, either plot or its data}
\usage{
makeSimpleHist(df, var, bins, makeTable = FALSE)
}
\arguments{
\item{df}{data frame from which output is to be made}

\item{var}{string defining which variable in the data frame to use}

\item{bins}{numeric vector defining the number of equally large groups}

\item{makeTable}{Logical that if TRUE function will return a data frame
containing the bin borders and count within each bin}
}
\value{
a graphical object or data frame
}
\description{
Short demo on how to produce dynamic content in a shiny app at Rapporteket
}
\examples{
makeSimpleHist(df = mtcars, var = "mpg", bins = 5, makeTable = FALSE)
}
