% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_over_tid.R
\name{over_tid_plot}
\alias{over_tid_plot}
\title{Funksjon for å lage kolonne-figur for datasettet med meslinger}
\usage{
over_tid_plot(data, valg_region)
}
\arguments{
\item{data}{datasett som har vært gjennom over_tid_utvalg()}

\item{valg_region}{valg av region-visning gjort av bruker i ui-delen}
}
\value{
en "stacked columns"-figur
}
\description{
Funksjon for å lage kolonne-figur for datasettet med meslinger
}
