% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_over_tid.R
\name{over_tid_utvalg}
\alias{over_tid_utvalg}
\title{Funksjon for å gjøre utvalg basert på ui-valg om variabel og region}
\usage{
over_tid_utvalg(data, var, valg_region)
}
\arguments{
\item{data}{dataramme (her brukes data på meslinger)}

\item{var}{variabelen som velges av bruker i ui-delen}

\item{valg_region}{valg av region gjort av bruker i ui-delen}
}
\description{
Funksjon for å gjøre utvalg basert på ui-valg om variabel og region
}
