% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_pivot.R
\name{pivot_server}
\alias{pivot_server}
\title{Server logic for the pivot module}
\usage{
pivot_server(id, user)
}
\arguments{
\item{id}{Character string module namespace}

\item{user}{reactiveValues object containing user information}
}
\value{
Server logic for the pivot module
}
\description{
Server logic for the pivot module
}
