% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_pivot.R
\name{pivot_ui}
\alias{pivot_ui}
\title{UI for the pivot module}
\usage{
pivot_ui(id)
}
\arguments{
\item{id}{Character string module namespace}
}
\value{
An shiny app ui object
}
\description{
UI for the pivot module
}
