% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_plots.R
\name{plots_ui}
\alias{plots_ui}
\title{Shiny module providing GUI and server logic for the plot tab}
\usage{
plots_ui(id)
}
\arguments{
\item{id}{Character string module namespace}
}
\value{
An shiny app ui object
}
\description{
Shiny module providing GUI and server logic for the plot tab
}
