% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the Shiny Application}
\usage{
run_app(browser = FALSE, logAsJson = FALSE)
}
\arguments{
\item{browser}{Open app in browser window}

\item{logAsJson}{Log in json-format}
}
\value{
An object representing the app
}
\description{
Run the Shiny Application
}
