% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_samlerapport.R
\name{samlerapport_server}
\alias{samlerapport_server}
\title{Server logic for samlerapport}
\usage{
samlerapport_server(id)
}
\value{
A Shiny app server object
}
\description{
Server logic for samlerapport
}
